add show one order

+++ app/Http/Controllers/OrderController.php
@@ -17,7 +17,6 @@ class OrderController extends Controller
             'payment_method' => ['required', 'string', Rule::in(['CashSyriatel', 'CashMTN', 'Cash', 'Bank'])],
             'order_location' => 'required|string',
         ]);
-
         $product = Product::findOrFail($request->product_id);
 
         if ($product->quantity < $request->quantity) {
@@ -43,9 +42,23 @@ class OrderController extends Controller
 
     public function showUserOrders()
     {
-        $orders = auth()->user()->orders()->with(['product', 'store'])->get();
+//         $orders = auth()->user()->orders()->with(['product', 'store'])->first();
+        $orders = auth()->user()->orders()->get();
 
         return response()->json(['message' => 'ok', 'orders' => $orders], 200);
+    }
+
+    public function showUserOrder($id)
+    {
+        // $order = Order::where('id', $id)->with(['product', 'store'])->first();
+        $order = Order::where('id', $id)->first();
+
+        if ($order) {
+            if (auth()->user()->id != $order->user_id) {
+                return response()->json(['message' => 'Permission denied.'], 403);
+            }
+            return response()->json(['message' => 'ok', 'order' => $order], 200);
+        } else return response()->json(['message' => 'Order not found.'], 404);
     }
 
     public function updateOrder(Request $request, $order_id)
@@ -56,19 +69,20 @@ class OrderController extends Controller
             return response()->json(['message' => 'Permission denied.'], 403);
         }
 
+        if (auth()->user()->id != $order->user_id) {
+            return response()->json(['message' => 'Permission denied.'], 403);
+        }
         $request->validate([
             'quantity' => 'sometimes|integer|min:1',
             'payment_method' => ['sometimes', 'string', Rule::in(['CashSyriatel', 'CashMTN', 'Cash', 'Bank'])],
             'order_location' => 'sometimes|string',
         ]);
-
         $newQuantity = $request->input('quantity', $order->quantity);
         $quantityDifference = $newQuantity - $order->quantity;
 
         if ($quantityDifference > 0 && $order->product->quantity < $quantityDifference) {
             return response()->json(['message' => 'Not enough stock available'], 400);
         }
-
         $totalPrice = $order->product->price * $newQuantity;
         $order->update([
             'quantity' => $newQuantity,
@@ -76,25 +90,21 @@ class OrderController extends Controller
             'payment_method' => $request->input('payment_method', $order->payment_method),
             'order_location' => $request->input('order_location', $order->order_location),
         ]);
-
         $order->product->update(['quantity' => $order->product->quantity - $quantityDifference]);
-
         return response()->json(['message' => 'Order updated successfully', 'order' => $order], 200);
     }
 
     public function deleteOrder($order_id)
     {
-        $order = Order::findOrFail($order_id);
-
-        if (auth()->user()->id != $order->user_id) {
-            return response()->json(['message' => 'Permission denied.'], 403);
-        }
-
-        $order->product->increment('quantity', $order->quantity);
-
-        $order->delete();
-
-        return response()->json(['message' => 'Order deleted successfully'], 204);
+        $order = Order::where('id', $order_id)->first();
+        if ($order) {
+            if (auth()->user()->id != $order->user_id) {
+                return response()->json(['message' => 'Permission denied.'], 403);
+            }
+            $order->product->increment('quantity', $order->quantity);
+            $order->delete();
+            return response()->json(['message' => 'Order deleted successfully'], 204);
+        } else return response()->json(['message' => 'Order not found.'], 404);
     }
